Fix SGCT 2.5 and GLM issues
The new verisons of SGCT and GLM caused some issues. Shader handles are no
longer used in SGCT 2.5, instead names of the shaders are used.

In GLM, transformations accepting 3 float arguments do no longer exsist.
Instead, a combination of mat4 and vec3 are used. An example:
`glm::scale(a, b, c)` is now written
`glm::scale(glm::mat4(1.0f), glm::vec3(a, b, c))`.

This also makes sure that the CoreVideo Apple framework is used so that the
program can be compiled on OS X.